service/nwm_uds: Add NetworkStatusChangeReason
This is a guess based on HW tests performed on real 3DSes. No reverse engineering was done because I was not really sure how to.

The following is a summary of the test results, for future reference:

This value is 0x0 initially.
After creating/connecting to a network, this value changes to 0x1. Afterwards, sending/receiving data does not change the value. Clients connecting or disconnecting does not change the value.
Failed attempts to connect do not change the value (0x0).

When the application destroys/disconnects from the network actively, this value does not change (0x1), and nor does the status value (this is pretty strange).
When the connection was lost due to the host closing the network or ejecting the client, this value changes to 0x4. However, on Citra, apparently these types of 'unexpected connection loss' are not handled yet.

On my 3DS, UDS won't let me recreate or reconnect unless I shutdown UDS. (Not sure if something is wrong with this.) When UDS is reinitialized, this value is set back to 0x0.

This should fix multiplayer for a few games, including Pokemon Mystery Dungeon: Gates to Infinity and likely more.