Add descriptions for camera and microphone usage on macos
Starting from macOS 10.14, users need to grant permissions for applications to use media capture (camera and microphone).

The application *must* provide a description in its Info.plist or it will be terminated due to privacy violation upon attempting to initialize camera/microphone.

Ideally these two strings should be translated, but since this is not really compatible with Qt's localization system and likely requires separate configuration, I'm currently not that interested in messing with it.